requirements.txt
python-telegram-bot~=22.1
pytz
apscheduler==3.10.4